api/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
langchain==0.3.18
langchain-google-genai==2.0.10
langchain-groq==0.2.4
langgraph==0.2.73
supabase==2.13.0
python-dotenv==1.0.1
tavily-python==0.5.0
beautifulsoup4==4.13.3
requests==2.32.3
pydantic==2.10.6
langchain-community==0.3.17
sendgrid==6.11.0
